requirements.txt
python-telegram-bot==21.5
openai
sqlite3
python-dotenv
